feat: extract token to separate contract
