+++ Clarinet.toml
@@ -5,9 +5,14 @@ name = "citycoin"
 path = "contracts/clarinet/sip-10-ft-standard.clar"
 depends_on = []
 
+[contracts.token]
+path = "contracts/clarinet/token.clar"
+depends_on = ["sip-010-trait"]
+
+
 [contracts.citycoin]
 path = "contracts/clarinet/citycoin.clar"
-depends_on = ["sip-010-trait"]
+depends_on = ["token"]
 
 # contracts listed below are used only in test suite
 [contracts.malicious]
