+++ tests/citycoin_test.ts
@@ -21,11 +21,13 @@ import {
   SPLIT_STACKER_PERCENTAGE,
   SPLIT_CITY_PERCENTAGE
 } from "../src/citycoin-client.ts"
+import { TokenClient } from "../src/token-client.ts";
 
 describe('[CityCoin]', () => {
   let chain: Chain;
   let accounts: Map<string, Account>;
   let client: CityCoinClient;
+  let tokenClient: TokenClient;
   let deployer: Account;
   let wallet_1: Account;
   let wallet_2: Account;
@@ -57,221 +59,11 @@ describe('[CityCoin]', () => {
     wallet_5 = accounts.get('wallet_5')!;
     wallet_6 = accounts.get('wallet_6')!;
 
-    client = new CityCoinClient(chain, deployer);
+    client = new CityCoinClient("citycoin", chain, deployer);
+    tokenClient = new TokenClient("token", chain, deployer);
   }
 
-  describe("SIP-010:", () => {
-    setupCleanEnv();
-
-    describe("transfer()", () => {
-      beforeEach(() => {
-        setupCleanEnv();
-      })
-
-      it("succeeds with no memo supplied", () => {
-        const from = wallet_1;
-        const to = wallet_2;
-        const amount = 100;
-
-        chain.mineBlock([
-          client.ftMint(amount, wallet_1)
-        ]);
-
-        const block = chain.mineBlock([
-          client.transfer(amount, from, to, from)
-        ]);
-
-        assertEquals(block.receipts.length, 1);
-        block.receipts[0].result.expectOk();
-        block.receipts[0].events.expectFungibleTokenTransferEvent(
-          amount,
-          from.address,
-          to.address,
-          'citycoins'
-        );
-      });
-
-      it("succeeds with memo supplied", () => {
-        const from = wallet_1;
-        const to = wallet_2;
-        const amount = 100;
-        const memo = new TextEncoder().encode("MiamiCoin is the first CityCoin");
-
-        chain.mineBlock([
-          client.ftMint(amount, wallet_1)
-        ]);
-
-        const block = chain.mineBlock([
-          client.transfer(amount, from, to, from, memo)
-        ]);
-
-        assertEquals(block.receipts.length, 1);
-        block.receipts[0].result.expectOk();
-
-        const expectedEvent = {
-          type: "contract_event", 
-          contract_event: {
-            contract_identifier: client.getContractAddress(),
-            topic: "print",
-            value: types.some(types.buff(memo))
-          }
-        }
-
-        const receipt = block.receipts[0];
-        assertEquals(receipt.events.length, 2);
-        assertEquals(receipt.events[0], expectedEvent);
-        receipt.events.expectFungibleTokenTransferEvent(
-          amount,
-          from.address,
-          to.address,
-          'citycoins'
-        );
-      });
-
-      it("fails with u1 when sender does not have enough funds", () => {
-        const from = wallet_1;
-        const to = wallet_2;
-
-        const block = chain.mineBlock([
-          client.transfer(100, from, to, from)
-        ]);
-
-        assertEquals(block.receipts.length, 1);
-        block.receipts[0].result.expectErr().expectUint(1);
-      });
-
-      it("fails with u2 when sender and recipient are the same", () => {
-        const from = wallet_1;
-        const to = wallet_1;
-
-        chain.mineBlock([
-          client.ftMint(100, from)
-        ])
-
-        const block = chain.mineBlock([
-          client.transfer(100, from, to, from)
-        ]);
-
-        assertEquals(block.receipts.length, 1);
-        block.receipts[0].result.expectErr().expectUint(2);
-      });
-
-      it("fails with u3 when token sender is different than transaction sender", () => {
-        const from = wallet_1;
-        const to = wallet_2;
-
-        const block = chain.mineBlock([
-          client.transfer(10, from, to, to)
-        ]);
-
-        assertEquals(block.receipts.length, 1);
-        block.receipts[0].result.expectErr().expectUint(3);
-      });
-    });
-
-    describe("get-name()", () => {
-      it("returns 'citycoins'", () => {
-        const result = client.getName().result;
-
-        result.expectOk().expectAscii("citycoins");
-      });
-    });
-
-    describe("get-symbol()", () => {
-      it("returns 'CYCN'", () => {
-        const result = client.getSymbol().result;
-
-        result.expectOk().expectAscii("CYCN");
-      });
-    });
-
-    describe("get-decimals()", () => {
-      it("returns 0", () => {
-        const result = client.getDecimals().result;
-
-        result.expectOk().expectUint(0);
-      });
-    });
-
-    describe("get-balance()", () => {
-      it("returns 0 when no tokens are minted", () => {
-        const result = client.getBalance(wallet_1).result;
-
-        result.expectOk().expectUint(0);
-      });
-
-      it("returns 100 after 100 tokens are minted to a wallet", () => {
-        chain.mineBlock([
-          client.ftMint(100, wallet_1)
-        ]);
-
-        const result = client.getBalance(wallet_1).result;
-
-        result.expectOk().expectUint(100);
-      });
-    });
-
-    describe("get-total-supply()", () => {
-      beforeEach(() => {
-        setupCleanEnv();
-      })
-
-      it("returns 0 when no tokens are minted", () => {
-        const result = client.getTotalSupply().result;
-
-        result.expectOk().expectUint(0);
-      });
-
-      it("returns 100 after 100 tokens are minted", () => {
-        chain.mineBlock([
-          client.ftMint(100, wallet_1)
-        ]);
-
-        const result = client.getTotalSupply().result;
-
-        result.expectOk().expectUint(100);
-      });
-
-      it("returns 250000 after a miner wins a block", () => {
-        // activate mining
-        chain.mineBlock([
-          client.setMiningActivationThreshold(1),
-          client.registerMiner(wallet_3)
-        ]);
-
-        // advance chain to block where mining is active
-        chain.mineEmptyBlock(MINING_ACTIVATION_DELAY);
-
-        // mine a block
-        const block = chain.mineBlock([
-          client.mineTokens(100, wallet_1)
-        ]);
-
-        // advance chain past miner reward window
-        chain.mineEmptyBlock(101);
-
-        // claim tokens so they are minted
-        chain.mineBlock([
-          client.claimTokenReward(block.height - 1, wallet_1)
-        ]);
-
-        const result = client.getTotalSupply().result;
-
-        result.expectOk().expectUint(250000);
-      });
-
-    });
-
-    describe("get-token-uri()", () => {
-      it("returns correct uri", () => {
-        const result = client.getTokenUri().result;
-        const tokenUri = "https://cdn.citycoins.co/metadata/citycoin.json";
-
-        console.log(`\n  URI: ${tokenUri}`)
-        result.expectOk().expectSome().expectUtf8(tokenUri);
-      });
-    });
-  });
+  
 
   describe("Read Only:", () => {
     setupCleanEnv();
@@ -315,7 +107,7 @@ describe('[CityCoin]', () => {
         const startStacksHeight = MINING_ACTIVATION_DELAY + 5;
 
         chain.mineBlock([
-          client.ftMint(100, wallet_1),
+          tokenClient.ftMint(100, wallet_1),
           client.stackTokens(100, startStacksHeight, 1, wallet_1)
         ]);
 
@@ -456,7 +248,7 @@ describe('[CityCoin]', () => {
 
         // stack in cycle 1 while cycle 0 is active
         chain.mineBlock([
-          client.ftMint(100, wallet_2),
+          tokenClient.ftMint(100, wallet_2),
           client.stackTokens(100, MINING_ACTIVATION_DELAY + 5, 1, wallet_2)
         ]);
 
@@ -516,7 +308,7 @@ describe('[CityCoin]', () => {
 
         // stack in cycle 1 while cycle 0 is active
         chain.mineBlock([
-          client.ftMint(100, wallet_2),
+          tokenClient.ftMint(100, wallet_2),
           client.stackTokens(100, MINING_ACTIVATION_DELAY + 5, 1, wallet_2)
         ]);
 
@@ -795,7 +587,7 @@ describe('[CityCoin]', () => {
         setupCleanEnv();
 
         chain.mineBlock([
-          client.ftMint(100, wallet_1),
+          tokenClient.ftMint(100, wallet_1),
           client.setMiningActivationThreshold(1),
           client.registerMiner(wallet_3)
         ]);
@@ -895,7 +687,7 @@ describe('[CityCoin]', () => {
 
         // add tokens and stack them at the next cycle
         chain.mineBlock([
-          client.ftMint(5000, stacker),
+          tokenClient.ftMint(5000, stacker),
           client.stackTokens(5000, 105, targetRewardCycle, stacker),
         ]);
 
@@ -928,7 +720,7 @@ describe('[CityCoin]', () => {
 
         // add tokens and stack them at the next cycle
         chain.mineBlock([
-          client.ftMint(5000, stacker),
+          tokenClient.ftMint(5000, stacker),
           client.stackTokens(5000, 105, targetRewardCycle, stacker),
         ]);
 
@@ -963,7 +755,7 @@ describe('[CityCoin]', () => {
 
         // add tokens and stack them at the next cycle
         chain.mineBlock([
-          client.ftMint(5000, stacker),
+          tokenClient.ftMint(5000, stacker),
           client.stackTokens(5000, MINING_ACTIVATION_DELAY + 5, targetRewardCycle, stacker),
         ]);
 
@@ -1001,8 +793,8 @@ describe('[CityCoin]', () => {
 
         // add tokens and stack them at the next cycle
         chain.mineBlock([
-          client.ftMint(stackerOneStacked, stackerOne),
-          client.ftMint(stackerTwoStacked, stackerTwo),
+          tokenClient.ftMint(stackerOneStacked, stackerOne),
+          tokenClient.ftMint(stackerTwoStacked, stackerTwo),
           client.stackTokens(stackerOneStacked, MINING_ACTIVATION_DELAY + 5, targetRewardCycle, stackerOne),
           client.stackTokens(stackerTwoStacked, MINING_ACTIVATION_DELAY + 5, targetRewardCycle, stackerTwo)
         ]);
@@ -1047,8 +839,8 @@ describe('[CityCoin]', () => {
 
         // add tokens and stack them at the next cycle
         chain.mineBlock([
-          client.ftMint(stackerOneStacked, stackerOne),
-          client.ftMint(stackerTwoStacked, stackerTwo),
+          tokenClient.ftMint(stackerOneStacked, stackerOne),
+          tokenClient.ftMint(stackerTwoStacked, stackerTwo),
           client.stackTokens(stackerOneStacked, MINING_ACTIVATION_DELAY + 5, targetRewardCycle, stackerOne),
           client.stackTokens(stackerTwoStacked, MINING_ACTIVATION_DELAY + 5, targetRewardCycle, stackerTwo)
         ]);
@@ -1146,7 +938,7 @@ describe('[CityCoin]', () => {
 
         // stack in cycle 1 while cycle 0 is active
         chain.mineBlock([
-          client.ftMint(100, wallet_2),
+          tokenClient.ftMint(100, wallet_2),
           client.stackTokens(100, 105, 1, wallet_2)
         ]);
 
@@ -1322,7 +1114,7 @@ describe('[CityCoin]', () => {
         chain.mineBlock([
           client.setMiningActivationThreshold(1),
           client.registerMiner(wallet_1),
-          client.ftMint(amount, wallet_1)
+          tokenClient.ftMint(amount, wallet_1)
         ]);
 
         chain.mineEmptyBlock(MINING_ACTIVATION_DELAY);
@@ -1362,7 +1154,7 @@ describe('[CityCoin]', () => {
         chain.mineBlock([
           client.setMiningActivationThreshold(1),
           client.registerMiner(wallet_1),
-          client.ftMint(tokensAmount, wallet_1)
+          tokenClient.ftMint(tokensAmount, wallet_1)
         ]);
 
         chain.mineEmptyBlock(MINING_ACTIVATION_DELAY);
@@ -1500,7 +1292,7 @@ describe('[CityCoin]', () => {
         const startStacksHeight = MINING_ACTIVATION_DELAY + 5;
 
         const block = chain.mineBlock([
-          client.ftMint(100, wallet_1),
+          tokenClient.ftMint(100, wallet_1),
           client.stackTokens(100, startStacksHeight, 1, wallet_1)
         ]);
 
@@ -1525,7 +1317,7 @@ describe('[CityCoin]', () => {
         const startStacksHeight = MINING_ACTIVATION_DELAY + 5;
 
         chain.mineBlock([
-          client.ftMint(1000, wallet_1),
+          tokenClient.ftMint(1000, wallet_1),
           client.stackTokens(100, startStacksHeight, 1, wallet_1)
         ]);
 
@@ -1632,7 +1424,7 @@ describe('[CityCoin]', () => {
         ])
 
         chain.mineBlock([
-          client.ftMint(100, wallet_1),
+          tokenClient.ftMint(100, wallet_1),
           client.stackTokens(100, startStacksHeight, 1, wallet_1)
         ]);
 
@@ -1715,7 +1507,7 @@ describe('[CityCoin]', () => {
 
         // add tokens and stack them at the next cycle
         chain.mineBlock([
-          client.ftMint(5000, stacker),
+          tokenClient.ftMint(5000, stacker),
           client.stackTokens(5000, 5, 1, stacker),
         ]);
 
@@ -1756,7 +1548,7 @@ describe('[CityCoin]', () => {
 
         // add tokens and stack them at the next cycle
         chain.mineBlock([
-          client.ftMint(stackedAmount, stacker),
+          tokenClient.ftMint(stackedAmount, stacker),
           client.stackTokens(stackedAmount, MINING_ACTIVATION_DELAY + 5, 1, stacker),
         ]);
 
@@ -1881,44 +1673,7 @@ describe('[CityCoin]', () => {
       });
     });
 
-    describe('set-token-uri', () => {
-      it("fails with ERR_UNAUTHORIZED when called by someone who is not contract owner", () => {
-        const block = chain.mineBlock([
-          client.setTokenUri(wallet_3, "http://something-something.com")
-        ]);
-
-        const receipt = block.receipts[0];
-
-        receipt.result.expectErr().expectUint(ErrCode.ERR_UNAUTHORIZED);
-      });
-
-      it("changes token uri to none if no new value is provided", () => {
-        const block = chain.mineBlock([
-          client.setTokenUri(deployer)
-        ]);
-
-        const receipt = block.receipts[0];
-        receipt.result.expectOk().expectBool(true);
-
-        const result = client.getTokenUri().result;
-        result.expectOk().expectNone();
-      });
-
-      it("changes token uri to new value if provided", () => {
-        const newUri = "http://something-something.com"
-        const block = chain.mineBlock([
-          client.setTokenUri(deployer, newUri)
-        ]);
-
-        const receipt = block.receipts[0];
-        receipt.result.expectOk().expectBool(true);
-
-        const result = client.getTokenUri().result;
-        result.expectOk().expectSome().expectUtf8(newUri);
-      });
-    });
-
-    describe("set-city-wallet", () => {
+      describe("set-city-wallet", () => {
       beforeEach(() => {
         setupCleanEnv();
       });
